Move redirect out of QR scanner component
Also removes duplicate component.
