+++ desktop/src/app/components/resources/resources.module.ts
@@ -50,7 +50,6 @@ import { Routing } from '../../services/routing';
 import { StandardStateSerializer } from '../../services/standard-state-serializer';
 import { StateSerializer } from '../../services/state-serializer';
 import { SettingsProvider } from '../../services/settings/settings-provider';
-import { QrCodeScannerModalComponent } from './searchbar/resources-search-modal-qr';
 
 const remote = typeof window !== 'undefined' ? window.require('@electron/remote') : undefined;
 
@@ -92,8 +91,7 @@ const remote = typeof window !== 'undefined' ? window.require('@electron/remote'
         MoveModalComponent,
         DeletionInProgressModalComponent,
         ChildrenViewComponent,
-        BaseList,
-        QrCodeScannerModalComponent
+        BaseList
     ],
     providers: [
         { provide: StateSerializer, useClass: StandardStateSerializer },
