+++ desktop/src/app/components/resources/searchbar/resources-search-constraints.component.ts
@@ -1,14 +1,15 @@
 import { clone } from 'tsfun';
 import { Component, Renderer2 } from '@angular/core';
 import { I18n } from '@ngx-translate/i18n-polyfill';
-import { Datastore, Field, ProjectConfiguration, Labels } from 'idai-field-core';
+import { Datastore, Field, ProjectConfiguration, Labels, FieldDocument } from 'idai-field-core';
 import { ViewFacade } from '../../../components/resources/view/view-facade';
 import { SearchConstraintsComponent } from '../../widgets/search-constraints.component';
 import { ResourcesSearchBarComponent } from './resources-search-bar.component';
-import { QrCodeScannerModalComponent } from './resources-search-modal-qr';
+import { QrCodeScannerModalComponent } from '../../widgets/resources-search-modal-qr';
 import { MenuContext } from '../../../services/menu-context';
 import { NgbModalRef, NgbModal } from '@ng-bootstrap/ng-bootstrap';
 import { Menus } from '../../../services/menus';
+import { Routing } from '../../../services/routing';
 
 
 
@@ -35,7 +36,8 @@ export class ResourcesSearchConstraintsComponent extends SearchConstraintsCompon
                 private viewFacade: ViewFacade,
                 labels: Labels,
                 private menus: Menus,
-                private modalService: NgbModal
+                private modalService: NgbModal,
+                private routingService: Routing
     ) {
 
         super(resourcesSearchBarComponent, projectConfiguration, datastore, renderer, labels, i18n);
@@ -140,12 +142,27 @@ export class ResourcesSearchConstraintsComponent extends SearchConstraintsCompon
                 { animation: false }
             );
 
-            //await modalRef.result;
+            this.openDocument(await modalRef.result);
             return true;
         } catch (_) {
             return false;
         } finally {
             this.menus.setContext(MenuContext.DEFAULT);
         }
+    }
+
+    
+    // to open the scanned document
+    private async openDocument(scannedCode: string) {
+        // split the scanned code with an '@'
+        console.log(scannedCode);
+
+        const [uuid, projectName] = scannedCode.split('@');
+        const document = (await this.datastore.get(uuid) as FieldDocument);
+
+        // open the scanned resource
+        this.routingService.jumpToResource(document);
+        // close the modal window
+        //this.activeModal.close();
     }
 }
