--- desktop/src/app/components/resources/searchbar/resources-search-modal-qr.html
@@ -1,25 +1,0 @@
-<h1
-    style="
-        text-align: center;
-        font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
-    "
->
-    Scanner
-</h1>
-<div>
-    <video style="max-width: 100%"></video>
-</div>
-
-<h2 *ngIf="!hasCamera" style="text-align: center; color: red;">
-    <span
-        class="mdi mdi-emoticon-sad-outline"
-        style="
-            font-size: xxx-large;
-            position: absolute;
-            left: 225px;
-            bottom: 150px;
-            color: red;
-        "
-    ></span>
-    Sorry <br> Camera not found
-</h2>
